+++ Editor/BacktraceConfigurationLabels.cs
@@ -7,16 +7,9 @@
         internal static string LABEL_REPORT_PER_MIN = "Reports per minute";
         internal static string LABEL_HANDLE_UNHANDLED_EXCEPTION = "Handle unhandled exceptions";
 
-#if UNITY_2018_4_OR_NEWER
         internal static string LABEL_IGNORE_SSL_VALIDATION = "Ignore SSL validation";
-#endif
-#if UNITY_ANDROID
         internal static string LABEL_HANDLE_ANR = "Handle ANR (Application not responding)";
-#endif
-#if UNITY_STANDALONE_WIN
         internal static string LABEL_SEND_UNHANDLED_GAME_CRASHES_ON_STARTUP= "Send unhandled native game crashes on startup";
-        internal static string LABEL_MINIDUMP_SUPPORT = "Minidump type";
-#endif
         internal static string LABEL_USE_NORMALIZED_EXCEPTION_MESSAGE = "Use normalized exception message";
         internal static string LABEL_GAME_OBJECT_DEPTH = "Game object depth limit";
         internal static string LABEL_DESTROY_CLIENT_ON_SCENE_LOAD = "Destroy client on new scene load (false - Backtrace managed)";
@@ -24,6 +17,8 @@
         // database labels
         internal static string LABEL_ENABLE_DATABASE = "Enable Database";
         internal static string LABEL_PATH = "Backtrace database path";
+        internal static string LABEL_MINIDUMP_SUPPORT = "Minidump type";
+        internal static string LABEL_ADD_UNITY_LOG = "Add Unity player log to report";
         internal static string LABEL_GENERATE_SCREENSHOT_ON_EXCEPTION = "Generate Screenshot when exception happened";
         internal static string LABEL_DEDUPLICATION_RULES = "Client-Side deduplication";
         internal static string LABEL_AUTO_SEND_MODE = "Auto send mode";