+++ Runtime/Model/Database/BacktraceDatabaseSettings.cs
@@ -8,37 +8,36 @@ namespace Backtrace.Unity.Model.Database
     /// </summary>
     public class BacktraceDatabaseSettings
     {
+        private readonly BacktraceConfiguration _configuration;
         public BacktraceDatabaseSettings(BacktraceConfiguration configuration)
         {
             if (configuration == null)
             {
                 return;
             }
-
-            DatabasePath = configuration.DatabasePath;
-            MaxRecordCount = Convert.ToUInt32(configuration.MaxRecordCount);
-            MaxDatabaseSize = configuration.MaxDatabaseSize;
-            AutoSendMode = configuration.AutoSendMode;
-            RetryInterval = Convert.ToUInt32(configuration.RetryInterval);
-            RetryLimit = Convert.ToUInt32(configuration.RetryLimit);
-            RetryOrder = configuration.RetryOrder;
-            DeduplicationStrategy = configuration.DeduplicationStrategy;
-            GenerateScreenshotOnException = configuration.GenerateScreenshotOnException;
+            _configuration = configuration;
         }
         /// <summary>
         /// Directory path where reports and minidumps are stored
         /// </summary>
-        public string DatabasePath { get; private set; }
+        public string DatabasePath
+        {
+            get
+            {
+                return _configuration.DatabasePath;
+            }
+        }
 
         /// <summary>
         /// Maximum number of stored reports in Database. If value is equal to zero, then limit not exists
         /// </summary>
-        public uint MaxRecordCount { get; set; }
-
-        /// <summary>
-        /// Database size in MB
-        /// </summary>
-        private long _maxDatabaseSize = 0;
+        public uint MaxRecordCount
+        {
+            get
+            {
+                return Convert.ToUInt32(_configuration.MaxRecordCount);
+            }
+        }
 
         /// <summary>
         /// Maximum database size in MB. If value is equal to zero, then size is unlimited
@@ -48,44 +47,96 @@ namespace Backtrace.Unity.Model.Database
             get
             {
                 //convert megabyte to bytes
-                return _maxDatabaseSize * 1000 * 1000;
-            }
-            set
-            {
-                _maxDatabaseSize = value;
+                return _configuration.MaxDatabaseSize * 1000 * 1000;
             }
         }
 
         /// <summary>
         /// Resend report when http client throw exception
         /// </summary>
-        public bool AutoSendMode { get; set; }
-
-        /// <summary>
-        /// Retry behaviour
-        /// </summary>
-        public RetryBehavior RetryBehavior = RetryBehavior.ByInterval;
+        public bool AutoSendMode
+        {
+            get
+            {
+                return _configuration.AutoSendMode;
+            }
+        }
 
         /// <summary>
         /// How much seconds library should wait before next retry.
         /// </summary>
-        public uint RetryInterval = 5;
+        public uint RetryInterval
+        {
+            get
+            {
+                return Convert.ToUInt32(_configuration.RetryInterval);
+            }
+        }
 
         /// <summary>
         /// Maximum number of retries
         /// </summary>
-        public uint RetryLimit = 3;
+        public uint RetryLimit
+        {
+            get
+            {
+                return Convert.ToUInt32(_configuration.RetryLimit);
+            }
+        }
 
         /// <summary>
         /// Deduplication strategy
         /// </summary>
-        public DeduplicationStrategy DeduplicationStrategy = DeduplicationStrategy.None;
+        public DeduplicationStrategy DeduplicationStrategy
+        {
+            get
+            {
+                return _configuration.DeduplicationStrategy;
+            }
+        }
 
         /// <summary>
         /// Generate screenshot on exception
         /// </summary>
-        public bool GenerateScreenshotOnException;
+        public bool GenerateScreenshotOnException
+        {
+            get
+            {
+                return _configuration.GenerateScreenshotOnException;
+            }
+        }
 
-        public RetryOrder RetryOrder = RetryOrder.Queue;
+        /// <summary>
+        /// Add Unity log file to Backtrace report
+        /// </summary>
+        public bool AddUnityLogToReport
+        {
+            get
+            {
+                return _configuration.AddUnityLogToReport;
+            }
+        }
+
+        public RetryOrder RetryOrder
+        {
+            get
+            {
+                return _configuration.RetryOrder;
+            }
+        }
+
+        public MiniDumpType MinidumpType
+        {
+            get
+            {
+
+#if UNITY_STANDALONE_WIN
+                return _configuration.MinidumpType;
+#else
+                return MiniDumpType.None;
+
+#endif
+            }
+        }
     }
 }