+++ Runtime/Services/BacktraceDatabaseContext.cs
@@ -5,15 +5,17 @@ using Backtrace.Unity.Model.Database;
 using Backtrace.Unity.Types;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Backtrace.Unity.Tests.Runtime")]
 namespace Backtrace.Unity.Services
 {
     /// <summary>
     /// Backtrace Database Context
     /// </summary>
-    public class BacktraceDatabaseContext : IBacktraceDatabaseContext
+    internal class BacktraceDatabaseContext : IBacktraceDatabaseContext
     {
         /// <summary>
         /// Database cache
@@ -40,8 +42,6 @@ namespace Backtrace.Unity.Services
         /// </summary>
         private readonly int _retryNumber;
 
-        private readonly bool _generateScreenshotOnException;
-
         /// <summary>
         /// Record order
         /// </summary>
@@ -52,35 +52,23 @@ namespace Backtrace.Unity.Services
         /// </summary>
         public DeduplicationStrategy DeduplicationStrategy { get; set; }
 
+        private readonly BacktraceDatabaseAttachmentManager _attachmentManager;
+
 
         /// <summary>
         /// Initialize new instance of Backtrace Database Context
         /// </summary>
         /// <param name="settings">Database settings</param>
         public BacktraceDatabaseContext(BacktraceDatabaseSettings settings)
-            : this(settings.DatabasePath, settings.RetryLimit, settings.RetryOrder, settings.DeduplicationStrategy)
-        { }
-
-        /// <summary>
-        /// Initialize new instance of Backtrace Database Context
-        /// </summary>
-        /// <param name="path">Path to database directory</param>
-        /// <param name="retryNumber">Total number of retries</param>
-        /// <param name="retryOrder">Record order</param>
-        public BacktraceDatabaseContext(
-           string path,
-           uint retryNumber,
-           RetryOrder retryOrder,
-           DeduplicationStrategy deduplicationStrategy = DeduplicationStrategy.None,
-           bool generateScreenshotOnException = false)
         {
-            _path = path;
-            _retryNumber = checked((int)retryNumber);
-            _generateScreenshotOnException = generateScreenshotOnException;
-            RetryOrder = retryOrder;
-            DeduplicationStrategy = deduplicationStrategy;
+            _path = settings.DatabasePath;
+            _retryNumber = checked((int)settings.RetryLimit);
+            _attachmentManager = new BacktraceDatabaseAttachmentManager(settings);
+            RetryOrder = settings.RetryOrder;
+            DeduplicationStrategy = settings.DeduplicationStrategy;
             SetupBatch();
         }
+
 
         /// <summary>
         /// Setup cache 
@@ -123,7 +111,7 @@ namespace Backtrace.Unity.Services
         /// </summary>
         /// <param name="backtraceData">Diagnostic data that should be stored in database</param>
         /// <returns>New instance of DatabaseRecordy</returns>
-        public BacktraceDatabaseRecord Add(BacktraceData backtraceData, MiniDumpType miniDumpType = MiniDumpType.None)
+        public BacktraceDatabaseRecord Add(BacktraceData backtraceData)
         {
             if (backtraceData == null)
             {
@@ -144,23 +132,15 @@ namespace Backtrace.Unity.Services
                     return existRecord;
                 }
             }
-
-            if (_generateScreenshotOnException)
+            //add built-in attachments
+            var attachments = _attachmentManager.GetReportAttachments(backtraceData);
+            foreach (var attachment in attachments)
             {
-                var screenshotPath = Path.Combine(_path, string.Format("{0}.png", backtraceData.Uuid));
-                UnityEngine.ScreenCapture.CaptureScreenshot(screenshotPath);
-                backtraceData.Attachments.Add(screenshotPath);
-                if (backtraceData.Report.AttachmentPaths != null)
+                backtraceData.Report.SetMinidumpPath(attachment);
+                if (!string.IsNullOrEmpty(attachment))
                 {
-                    backtraceData.Report.AttachmentPaths.Add(screenshotPath);
+                    backtraceData.Attachments.Add(attachment);
                 }
-            }
-
-            string minidumpPath = GenerateMiniDump(backtraceData.Report, miniDumpType);
-            backtraceData.Report.SetMinidumpPath(minidumpPath);
-            if (!string.IsNullOrEmpty(minidumpPath))
-            {
-                backtraceData.Attachments.Add(minidumpPath);
             }
 
             var record = ConvertToRecord(backtraceData, hash);
@@ -262,7 +242,6 @@ namespace Backtrace.Unity.Services
                     }
                 }
             }
-            return;
         }
 
         /// <summary>