+++ Tests/Runtime/Mocks/BacktraceDatabaseContextMock.cs
@@ -2,20 +2,23 @@
 using Backtrace.Unity.Model.Database;
 using Backtrace.Unity.Services;
 
-public class BacktraceDatabaseContextMock : BacktraceDatabaseContext
+namespace Backtrace.Unity.Tests.Runtime
 {
-    private BacktraceDatabaseSettings _settings;
-    public BacktraceDatabaseContextMock(BacktraceDatabaseSettings settings) : base(settings)
+    internal class BacktraceDatabaseContextMock : BacktraceDatabaseContext
     {
-        _settings = settings;
-    }
-
-    protected override BacktraceDatabaseRecord ConvertToRecord(BacktraceData backtraceData, string hash)
-    {
-        //create new record and return it to AVOID storing data on hard drive
-        return new BacktraceDatabaseRecord(backtraceData, _settings.DatabasePath)
+        private BacktraceDatabaseSettings _settings;
+        public BacktraceDatabaseContextMock(BacktraceDatabaseSettings settings) : base(settings)
         {
-            Hash = hash
-        };
+            _settings = settings;
+        }
+
+        protected override BacktraceDatabaseRecord ConvertToRecord(BacktraceData backtraceData, string hash)
+        {
+            //create new record and return it to AVOID storing data on hard drive
+            return new BacktraceDatabaseRecord(backtraceData, _settings.DatabasePath)
+            {
+                Hash = hash
+            };
+        }
     }
 }